+++ crates/tytanic/src/cli/commands/util/clean.rs
@@ -5,6 +5,7 @@ use termcolor::Color;
 use tytanic_utils::fmt::Term;
 
 use super::Context;
+use crate::cli::commands::FilterOptions;
 use crate::cwrite;
 
 #[derive(clap::Args, Debug, Clone)]
@@ -13,15 +14,18 @@ pub struct Args {
     /// Also remove persistent references
     #[arg(long)]
     pub include_persistent_references: bool,
+
+    #[command(flatten)]
+    pub filter: FilterOptions,
 }
 
 pub fn run(ctx: &mut Context, args: &Args) -> eyre::Result<()> {
     let project = ctx.project()?;
-    let suite = ctx.collect_tests(&project)?;
+    let suite = ctx.collect_tests_with_filter(&project, ctx.filter(&args.filter)?)?;
 
     let mut temp = 0;
     let mut persistent = 0;
-    for test in suite.unit_tests() {
+    for test in suite.matched().unit_tests() {
         test.delete_temporary_directories(&project)?;
         if args.include_persistent_references && test.kind().is_persistent() {
             test.delete_reference_document(&project)?;