+++ docs/CHANGELOG.md
@@ -14,6 +14,7 @@
 - Added `util manpage` sub command for generating man pages
 - Unhid non-default switches on short help
 - Added `--include-persistent-references` to `util clean` sub command
+- Added filter options to `util clean` sub command
 
 ## Fixes
 - Don't panic when trying to update non-persistent tests